+++ src/main/java/com/create/chacha/domains/seller/areas/reviews/service/SellerReviewStatsService.java
@@ -1,0 +1,8 @@
+package com.create.chacha.domains.seller.areas.reviews.service;
+
+import com.create.chacha.domains.seller.areas.reviews.dto.response.ReviewStatsResponseDTO;
+
+public interface SellerReviewStatsService {
+    ReviewStatsResponseDTO getStoreStats(String storeUrl);
+    ReviewStatsResponseDTO getProductStats(String storeUrl, Long productId);
+}
